1. 部分groupId原来使用的是int, 现统一为Long. 2. 上新2.0增加的一些内容(编码中, 对其他业务无影响)

+++ voyageone-com/src/main/java/com/voyageone/cms/service/model/CmsBtProductModel_Group.java
@@ -90,7 +90,7 @@ public class CmsBtProductModel_Group extends BaseMongoMap<String, Object> {
         setAttribute("platforms", platforms);
     }
 
-    public CmsBtProductModel_Group_Platform getPlatformByGroupId(int groupId) {
+    public CmsBtProductModel_Group_Platform getPlatformByGroupId(Long groupId) {
         CmsBtProductModel_Group_Platform result = null;
         List<CmsBtProductModel_Group_Platform> platforms = getPlatforms();
         for (CmsBtProductModel_Group_Platform platform : platforms) {
